Finished Project

+++ src/api/models/Bread.java
@@ -3,7 +3,7 @@ package api.models;
 
 public class Bread extends Product
 {
-    public Bread(String name, Type type, int price) {
+    public Bread(String name, Type type, double price) {
         super(name, type, price);
     }
 }
