+++ src/api/models/Eggs.java
@@ -3,7 +3,7 @@ package api.models;
 
 public class Eggs extends Product
 {
-    public Eggs(String name, Type type, int price) {
+    public Eggs(String name, Type type, double price) {
         super(name, type, price);
     }
 }
