+++ src/api/models/Shirt.java
@@ -3,7 +3,7 @@ package api.models;
 
 public class Shirt extends Product
 {
-    public Shirt(String name, Type type, int price) {
+    public Shirt(String name, Type type, double price) {
         super(name, type, price);
     }
 }
