+++ src/api/models/Trousers.java
@@ -3,7 +3,7 @@ package api.models;
 
 public class Trousers extends Product
 {
-    public Trousers(String name, Type type, int price) {
+    public Trousers(String name, Type type, double price) {
         super(name, type, price);
     }
 }
